Noderef transform framework.
Each noderef contains a linked list of transforms, which consist of a type,
a function pointer and arbitrary arguments (union of structs).

Among other things, this allows geometric transformations or filters to
be applied to node elements and saved with the maps, arguments included;
altered tiles can be cached for expensive transforms.