Noderef transform framework.
Each noderef contains a linked list of transforms, which consist of a type,
a function pointer and arbitrary arguments (union of structs).

Among other things, this allows geometric transformations or filters to
be applied to node elements and saved with the maps, arguments included;
altered tiles can be cached for expensive transforms.


git-svn-id: https://dev.csoft.net/agar/trunk@1898 86a84d31-9702-0410-8f1e-a60cab7bd424