Allow creation of a direct rendering context if possible bypassing the X server. If the display is remote or otherwise not accepting direct rendering the indirect rendering context is selected automatically.
From http://www.opengl.org/sdk/docs/man/xhtml/glXCreateContext.xml :

	GLXContext glXCreateContext(Display *  dpy,  XVisualInfo *  vis,
		GLXContext shareList,  Bool  direct);

"If direct is True, then a direct rendering context is created if the
implementation supports direct rendering, if the connection is to an X server
that is local, and if a direct rendering context is available. (An
implementation may return an indirect context when direct is True.) If direct
is False, then a rendering context that renders through the X server is always
created.  Direct rendering provides a performance advantage in some
implementations.  However, direct rendering contexts cannot be shared outside
a single process, and they may be unable to render to GLX pixmaps."

Since Agar is running on systems where SDL works, and SDL uses this flag i
think its safe to set.

XXX Maybe set it as an option?



git-svn-id: https://dev.csoft.net/agar/trunk@9006 86a84d31-9702-0410-8f1e-a60cab7bd424