Introduce the base framework for micro-Agar.
micro-Agar is a memory-constrained, 12-bit color version of the Agar GUI for
platforms such as the Z80, 6502 and 68000. It is not a port, but a separate
library entirely. However it still uses Agar-Core and the Agar object system.

Micro-Agar will be automatically selected by configure when cross-compiling
to supported targets (e.g., ./configure --host=c64).



git-svn-id: https://dev.csoft.net/agar/trunk@10495 86a84d31-9702-0410-8f1e-a60cab7bd424