Fix legibility and cosmetic issues with HSVPal. Closes issue #41.
Give visual feedback when controls are pressed. Preview color in the hue
selection circle for easier comparison. Highlight the controls with the
RGB inverse of the preview color.

Choose more sensibly the colors used by the RGB/HSV value labels.

Define AG_SATURATION_EPSILON and AG_VALUE_EPSILON and short-circuit the
cases where s or v are close to 0.



git-svn-id: https://dev.csoft.net/agar/trunk@10596 86a84d31-9702-0410-8f1e-a60cab7bd424