Handle word selections correctly in AG_Editable(3).
Instead of representing selections as a cursor position and signed offset (sel),
use two normalized positions selStart and selEnd independent from the cursor.

This enables us to fix the behavior of the incremental word-selection algorithm
whenever selections are reversed.  As a bonus it also reduces the amount of
computation draw() needs to do to determine whether a character is selected.

Simplify the shift-key selection-updating code (use a reference position).

Indicate selected newlines visually with a small vertical line.



git-svn-id: https://dev.csoft.net/agar/trunk@10777 86a84d31-9702-0410-8f1e-a60cab7bd424