AG_HSVPal: Fix incorrect handling of buttondown in NOALPHA case.
AG_HSVPal: Hue navigation with (Shift|Ctrl|Alt) + (Up|Down|Left|Right).

sdlfb, sdlgl: If (width, height) is not specified or "auto" default to (95%,90%)
as opposed to the previous 2/3 or 100% if display size is (<=700 or <=1000).

ag_math: Remove the <altivec.h> include. It is not required since we are passing
-faltivec to the linker.

Make Courier Prime the default font for AG_Console.



git-svn-id: https://dev.csoft.net/agar/trunk@10870 86a84d31-9702-0410-8f1e-a60cab7bd424