Offload AG_Window's fade{in,out} & opacity fields to a separate structure.
Create the AG_WindowFadeCtx on demand in order to reduce the size of the
AG_Window structure whenever fade and opacity features are not required.

Check "visible" & "dirty" flags locklessly in AG_WindowDrawQueued() and also
whenever evaluating focusable candidates after a focused window is hidden.

Prevent an unnecessary call to the driver's setWindowCaption() whenever a
window is shown initially and the caption does not need to be updated. Add
UPDATECAPTION flag to keep track of this.

Count the number of pinned windows in agWindowPinnedCount so we can optimize
the most common case (where the are no pinned windows to update).

Fix a cosmetic off-by-one when clearing BG after hiding windows in FB mode.

Document AG_WidgetUpdateCoords() and AG_ViewCapture().



git-svn-id: https://dev.csoft.net/agar/trunk@10880 86a84d31-9702-0410-8f1e-a60cab7bd424