Deliver `mouse-button-down` to multiple intersecting widgets.
If coordinates hit multiple widgets, deliver the event to all of them
as opposed to only the first match. This fixes a problem with widgets that
have been reordered, detached/reattached or hidden and re-shown (e.g.,
widgets under a Notebook).

Add build option DEBUG_MOUSE (debug delivery of mouse events to widgets).



git-svn-id: https://dev.csoft.net/agar/trunk@10933 86a84d31-9702-0410-8f1e-a60cab7bd424